+++ ColorCodedLog.py
@@ -2,32 +2,40 @@
  * File: ColorCodedLog.py
  * Copyright (c) 2023 Loupe
  * https://loupe.team
- * 
+ *
  * This file is part of ASPython, licensed under the MIT License.
 '''
-# Python Modules
+"""Backwards-compatibility shim.
+
+The old ``InitializeLogger`` / ``CustomFormatter`` API was unused inside the project, but is
+preserved here. New code should call ``aspython.logging_setup.setup_logging`` instead.
+"""
+import warnings as _warnings
+
+# The original implementation lived only in this module and was never imported by other
+# ASPython files; preserve verbatim under a sub-module name so external callers still work.
 import logging
 import sys
-from typing import Optional, Union 
+from typing import Optional, Union
 
-def InitializeLogger(logName:str="main", logLevel:Union[int, str, None]=None, logStringFormat:Optional[str]=None, infoColor:Optional[str]=None, debugColor:Optional[str]=None, warningColor:Optional[str]=None, errorColor:Optional[str]=None, criticalColor:Optional[str]=None, disableColors:Optional[bool]=False):
-    """Function for Initalizing a logger with a custom format"""
-    
-    # Find logger with given name. if it does not exist it will create one
+
+def InitializeLogger(logName: str = "main", logLevel: Union[int, str, None] = None,
+                     logStringFormat: Optional[str] = None, infoColor: Optional[str] = None,
+                     debugColor: Optional[str] = None, warningColor: Optional[str] = None,
+                     errorColor: Optional[str] = None, criticalColor: Optional[str] = None,
+                     disableColors: Optional[bool] = False):
     logger = logging.getLogger(logName)
+    if logLevel:
+        logger.setLevel(logLevel)
 
-    # Set logging level 
-    if logLevel: logger.setLevel(logLevel)
-
-    # For loop through handlers to see if one already exists
     customHandler = None
+    logFormatter = None
     for handler in logger.handlers:
         if isinstance(handler.formatter, CustomFormatter):
             customHandler = handler
             logFormatter = handler.formatter
             break
 
-    # Create custom handler if one does not exist
     if customHandler is None:
         customHandler = logging.StreamHandler(sys.stderr)
         customHandler.setLevel(logging.DEBUG)
@@ -35,10 +43,9 @@ def InitializeLogger(logName:str="main", logLevel:Union[int, str, None]=None, lo
         customHandler.setFormatter(logFormatter)
         logger.addHandler(customHandler)
 
- 
-    if logStringFormat: logFormatter.msgFormat = logStringFormat
+    if logStringFormat:
+        logFormatter.msgFormat = logStringFormat
 
-    # Disable colors by removing the color from the format string
     if disableColors:
         logFormatter.debug = ""
         logFormatter.info = ""
@@ -46,38 +53,39 @@ def InitializeLogger(logName:str="main", logLevel:Union[int, str, None]=None, lo
         logFormatter.error = ""
         logFormatter.critical = ""
     else:
-        # Only apply color if argument was passed in
-        if debugColor: logFormatter.debug = debugColor
-        if infoColor: logFormatter.info = debugColor
-        if warningColor: logFormatter.warning = debugColor
-        if errorColor: logFormatter.error = debugColor
-        if criticalColor: logFormatter.critical = debugColor
+        if debugColor:
+            logFormatter.debug = debugColor
+        if infoColor:
+            logFormatter.info = debugColor
+        if warningColor:
+            logFormatter.warning = debugColor
+        if errorColor:
+            logFormatter.error = debugColor
+        if criticalColor:
+            logFormatter.critical = debugColor
 
     return logger
 
-class CustomFormatter(logging.Formatter):
-    """Logging Formatter to add colors and count warning / errors"""
 
-    # Default values 
+class CustomFormatter(logging.Formatter):
     defaultDebug = "\x1b[38;21m"
     defaultInfo = "\x1b[38;21m"
     defaultWarning = "\x1b[33;21m"
     defaultError = "\x1b[31;21m"
     defaultCritical = "\x1b[31;1m"
-    defaultReset = "\x1b[0m"   
+    defaultReset = "\x1b[0m"
     defaultMsgFormat = "%(asctime)s - %(name)s - %(levelname)s - %(message)s (%(filename)s:%(lineno)d)"
 
     def __init__(self):
         self._debug = CustomFormatter.defaultDebug
         self._info = CustomFormatter.defaultInfo
-        self._warning = CustomFormatter.defaultWarning 
+        self._warning = CustomFormatter.defaultWarning
         self._error = CustomFormatter.defaultError
         self._critical = CustomFormatter.defaultCritical
         self._msgFormat = CustomFormatter.defaultMsgFormat
         self._reset = CustomFormatter.defaultReset
         self.generate()
 
-    # Make setter regenterate formatDict for each property 
     @property
     def debug(self):
         return self._debug
@@ -86,7 +94,6 @@ class CustomFormatter(logging.Formatter):
     def debug(self, value):
         self._debug = value
         self.generate()
-        return self._debug
 
     @property
     def info(self):
@@ -96,7 +103,6 @@ class CustomFormatter(logging.Formatter):
     def info(self, value):
         self._info = value
         self.generate()
-        return self._info
 
     @property
     def warning(self):
@@ -106,7 +112,6 @@ class CustomFormatter(logging.Formatter):
     def warning(self, value):
         self._warning = value
         self.generate()
-        return self._warning
 
     @property
     def error(self):
@@ -116,7 +121,6 @@ class CustomFormatter(logging.Formatter):
     def error(self, value):
         self._error = value
         self.generate()
-        return self._error
 
     @property
     def critical(self):
@@ -126,7 +130,6 @@ class CustomFormatter(logging.Formatter):
     def critical(self, value):
         self._critical = value
         self.generate()
-        return self._critical
 
     @property
     def msgFormat(self):
@@ -136,7 +139,6 @@ class CustomFormatter(logging.Formatter):
     def msgFormat(self, value):
         self._msgFormat = value
         self.generate()
-        return self._msgFormat
 
     @property
     def reset(self):
@@ -146,7 +148,6 @@ class CustomFormatter(logging.Formatter):
     def reset(self, value):
         self._reset = value
         self.generate()
-        return self._reset
 
     def generate(self):
         self._formatDict = {
@@ -154,36 +155,19 @@ class CustomFormatter(logging.Formatter):
             logging.INFO: self.info + self.msgFormat + self.reset,
             logging.WARNING: self.warning + self.msgFormat + self.reset,
             logging.ERROR: self.error + self.msgFormat + self.reset,
-            logging.CRITICAL: self.critical + self.msgFormat + self.reset
+            logging.CRITICAL: self.critical + self.msgFormat + self.reset,
         }
         return self
 
     def format(self, record):
         log_fmt = self._formatDict.get(record.levelno)
-        formatter = logging.Formatter(log_fmt)
-        return formatter.format(record)
-
-def main():
-
-    logger = InitializeLogger(logLevel="DEBUG")
-    logger.info('info message')
-    logger.debug('debug message')
-    logger.warning('warning message')
-    logger.error('error message')
-
-    logger2 = InitializeLogger(debugColor="\x1b[31;21m")
-    logger2.debug('debug message red')
-
-    logger = InitializeLogger(disableColors=True)
-    logger.info('info message')
-    logger.debug('debug message')
-    logger.warning('warning message')
-    logger.error('error message')
-    
-    logger2 = InitializeLogger(debugColor="\x1b[31;21m")
-    logger2.debug('debug message red')
+        return logging.Formatter(log_fmt).format(record)
 
 
-if __name__ == "__main__":
-    main()
-    
+_warnings.warn(
+    "Importing from 'ColorCodedLog' is deprecated; use 'aspython.logging_setup.setup_logging' instead.",
+    DeprecationWarning,
+    stacklevel=2,
+)
+
+__all__ = ["InitializeLogger", "CustomFormatter"]